re-order components
