+++ src/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import './App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Books from './redux/books/books';
-import Categories from './redux/categories/categories';
-import Form from './components/form';
+import Books from './components/books/books';
+import Categories from './components/categories';
+import Form from './components/books/form';
 // eslint-disable-next-line import/extensions
 import NavBar from './components/navBar';
