+++ src/components/books/booksReducer.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const booksReducer = () => (
+  <div />
+);
+
+export default booksReducer;
