+++ src/components/categories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Categories = () => (
+  <div>
+    <p>Under construction</p>
+  </div>
+);
+
+export default Categories;
