--- src/redux/categories/categories.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Categories = () => (
-  <div>
-    <p>Under construction</p>
-  </div>
-);
-
-export default Categories;
